Start working on a new editor
Based on the ideas in [write][], which itself is inspired by iA Writer
and uses ProseMirror.  This iteration uses CodeMirror, which is *much*
more appropriate, and makes writing a lot more fun because it simply
displays the text, and doesn't have as many kinks as ProseMirror has.

(The mismatch with ProseMirror was due to the fact that I wanted to
write *markdown*, but ProseMirror is for *fancy* text, not markdown.
For some reason I didn't think of CodeMirror until Marijn Haverbeke
pointed out to me that using it would be likely much easier.)