Change deps target to build dependencies correctly
The problem is that this project isn't developed inside a "proper" go
environment, but a directory, with the GOPATH being `.go`, a
subdirectory.  This means regular `go get -v ./...` works, but we have
to work around it by listing the dependencies manually, and then
running `go get` against that list.